+++ src/store/firebase.js
@@ -1,0 +1,16 @@
+import firebase from 'firebase/app';
+import 'firebase/auth'
+
+const config = {
+	apiKey: "AIzaSyBTtRuNAeLj9RdkYz5XpsC7u579QqLAmpg",
+    authDomain: "locasearch.firebaseapp.com",
+    databaseURL: "https://locasearch.firebaseio.com",
+    projectId: "locasearch",
+    storageBucket: "locasearch.appspot.com",
+    messagingSenderId: "660314635299"
+};
+
+export const firebaseapp = firebase.initializeApp(config);
+
+
+
